refactor: extrai o quiz para quiz.service
Move a correção do quiz (corrige no servidor, registra as respostas, conclui a aula e dispara a premiação) e a conferência de resposta única para quiz.service. O mínimo de acertos passa a ser exportado dali e reusado pelo estúdio. Controller em 868 linhas. Comportamento preservado.
